remove grouped rows
  - grouped rows was not used by any WebDriver tests and it has no special function
  - remove link, page, route, controller and mixin, model
  - removed Mountebank setup for model 'report'
