+++ app/mixins/chunked-group-data-mixin.js
@@ -7,19 +7,17 @@ export default Ember.Mixin.create({
 
   model: function () {
     var self = this;
+    var groupingMetadata = this.get('groupingMetadata');
     var tableContent = LazyGroupRowArray.create({
       loadChildren: function (chunkIndex, parentQuery) {
-        var query = {
-          chunkedGroup: 1,
-          section: chunkIndex + 1
-        };
+        var params = {content: parentQuery, groupingMetadata: groupingMetadata};
+        Ember.setProperties(params.content, {section: chunkIndex + 1});
         var sortName = self.get('sortName');
         if(sortName){
-          query.sortDirect = self.get('sortDirect');
-          query.sortName = sortName;
+          params.content.sortDirect = self.get('sortDirect');
+          params.content.sortName = sortName;
         }
-        Ember.setProperties(query, parentQuery);
-        return self.store.find('chunked-group', query).then(function (result) {
+        return self.store.find('chunked-group', params).then(function (result) {
           var meta = self.store.metadataFor('chunked-group');
           return {
             content: result,
@@ -30,12 +28,10 @@ export default Ember.Mixin.create({
           };
         });
       },
-      groupingMetadata: self.get('groupingMetadata')
+      groupingMetadata: groupingMetadata
     });
     return tableContent;
   },
-
-  groupingMetadata: [{id: 'accountSection'}, {id: 'accountType'}, {id: 'glAccountCode'}],
 
   actions: {
     setSortConditions: function (column) {
@@ -43,5 +39,7 @@ export default Ember.Mixin.create({
       this.set('sortName', columnName);
       this.set('sortDirect', column.get('currentDirect'));
     }
-  }
+  },
+
+  groupingMetadata: [{id: 'accountSection'}, {id: 'accountType'}, {id: 'accountCode'}]
 });
