--- app/mixins/group-rows-mixin.js
@@ -1,7 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Mixin.create({
-  model: function () {
-    return [this.store.find('report', 1)];
-  }
-});
