+++ app/router.js
@@ -13,7 +13,6 @@ export default Router.map(function() {
   this.route('groupsSort',{path: '/groups-sort'});
   this.route('groupingColumn', {path: 'grouping-column'});
   this.route('groupingColumnAndFixed', {path: 'grouping-column-and-fixed'});
-  this.route('groupedRows', {path: 'grouped-rows'});
   this.route('customGroupIndicator', {path: 'custom-group-indicator'});
   this.route('chunkedGroupingRows', {path: 'chunked-grouping-rows'});
   this.route('groupedRowLoadingIndicator', {path: 'grouped-row-loading-indicator'});
