--- app/routes/grouped-rows.js
@@ -1,4 +1,0 @@
-import Ember from 'ember';
-import GroupRowsMixin from '../mixins/group-rows-mixin';
-
-export default Ember.Route.extend(GroupRowsMixin, {});
