+++ app/templates/chunked-grouping-rows.hbs
@@ -3,6 +3,7 @@
 <ul>
     <li>AccountSection</li>
     <li>AccountType</li>
+    <li>AccountCode</li>
 </ul>
 <div class="lazy-loaded-table-container" style="height:200px; padding: 10px;">
   {{ember-table
