--- app/templates/grouped-rows.hbs
@@ -1,9 +1,0 @@
-<h3 class="inline-header">Grouped Rows</h3>
-<div class="lazy-loaded-table-container" style="height:600px;">
-  {{ember-table
-	  hasFooter=false
-	  enableContentSelection=true
-	  content=tableContent
-	  columns=columns
-  }}
-</div>
