+++ app/templates/index.hbs
@@ -7,7 +7,6 @@
     <li>{{#link-to 'groupsSort'}}Column Groups Sort inner column{{/link-to}}</li>
     <li>{{#link-to 'groupingColumn'}}Grouping Column{{/link-to}}</li>
     <li>{{#link-to 'groupingColumnAndFixed'}}Grouping Column And Fixed Columns{{/link-to}}</li>
-    <li>{{#link-to 'groupedRows'}}Grouped rows{{/link-to}}</li>
     <li>{{#link-to 'customGroupIndicator'}}Custom Group Indicator{{/link-to}}</li>
     <li>{{#link-to 'chunkedGroupingRows'}}Chunked Grouping Rows{{/link-to}}</li>
     <li>{{#link-to 'groupedRowLoadingIndicator'}}Rows with loading indicator{{/link-to}}</li>
